+++ NEWS.md
@@ -1,7 +1,7 @@
 # News
 
 ## master
-- Allow arbitrary reversible scale functions through `Makie.ReversibleScale`.
+- Allow arbitrary reversible scale functions through `ReversibleScale`.
 
 - Fixed some errors around dynamic changes of `ax.xscale` or `ax.yscale` [#3084](https://github.com/MakieOrg/Makie.jl/pull/3084)
 - Improved Barplot Label Alignment [#3160](https://github.com/MakieOrg/Makie.jl/issues/3160).